+++ OpenPNM/Physics/__GenericPhysics__.py
@@ -84,7 +84,6 @@ class GenericPhysics(OpenPNM.Base.Core):
         # Convert self.name into 'all'
         if key.split('.')[-1] == self.name:
             key = element + '.all'
-
         if key in self.keys():  # Look for data on self...
             return super(GenericPhysics, self).__getitem__(key)
         else:  # ...Then check Network
