+++ test/integration/script_test.py
@@ -259,66 +259,3 @@ def test_Darcy_alg():
     a = round(sp.absolute(Darcy2.rate(inlets))[0], 16)
     b = round(sp.absolute(Darcy2.rate(outlets))[0], 16)
     assert a == b
-
-
-def test_mapping():
-    # Create small cubic network
-    pn = OpenPNM.Network.Cubic(shape=[3, 3, 3], spacing=0.0001)
-    # Assign 3 different geometries to each layer in the z-direction
-    Pa = sp.arange(0, 9)
-    Ta = pn.find_neighbor_throats(Pa)
-    geom1 = OpenPNM.Geometry.GenericGeometry(network=pn, pores=Pa, throats=Ta)
-    Pc = sp.arange(18, 27)
-    Tc = pn.find_neighbor_throats(Pc)
-    geom3 = OpenPNM.Geometry.GenericGeometry(network=pn, pores=Pc, throats=Tc)
-    Pb = sp.arange(9, 18)
-    Tb = pn.find_neighbor_throats(pores=Pb, mode='intersection')
-    geom2 = OpenPNM.Geometry.GenericGeometry(network=pn, pores=Pb, throats=Tb)
-    # Create an index in the Network
-    pn['pore.num1'] = pn.Ps
-    # Create the same index across each geom
-    geom1['pore.num2'] = Pa
-    geom2['pore.num2'] = Pb
-    geom3['pore.num2'] = Pc
-    # Confirm two indexes match
-    assert(sp.all(pn['pore.num1'] == pn['pore.num2']))
-    # Send junk pores to ensure error is raised
-    with pytest.raises(Exception):
-        pn.map_pores(pores=[0, pn.Np-1], target=geom1)
-        pn.map_pores(pores=[0, pn.Np+1], target=geom1)
-        pn.map_pores(pores=[pn.Np-1], target=geom1)
-        pn.map_pores(pores=[pn.Np+1], target=geom1)
-        geom1.map_pores(pores=[0, geom1.Np+1], target=pn)
-        geom1.map_pores(pores=[0, pn.Np+1], target=pn)
-        geom1.map_pores(pores=[geom1.Np+1], target=pn)
-        geom1.map_pores(pores=[pn.Np+1], target=pn)
-        geom2.map_pores(pores=[0], target=geom1)
-        geom2.map_pores(pores=[geom2.Np+1], target=geom1)
-        geom2.map_pores(pores=[0, geom2.Np-1], target=geom1)
-        geom2.map_pores(pores=[0, geom2.Np+1], target=geom1)
-    # Trim column from center of Network
-    pn.trim(pores=[4, 13, 22])
-    # Confirm index still match
-    assert(sp.all(pn['pore.num1'] == pn['pore.num2']))
-    # Check mapping between each Geometry object and in both directions
-    # Check geom1
-    a = geom1.map_pores(pores=geom1.Ps, target=pn)
-    b = pn.map_pores(pores=a, target=geom1)
-    assert(sp.all(b == geom1.Ps))
-    a = geom1.map_throats(throats=geom1.Ts, target=pn)
-    b = pn.map_throats(throats=a, target=geom1)
-    assert(sp.all(b == geom1.Ts))
-    # Check geom2
-    a = geom2.map_pores(pores=geom2.Ps, target=pn)
-    b = pn.map_pores(pores=a, target=geom2)
-    assert(sp.all(b == geom2.Ps))
-    a = geom2.map_throats(throats=geom2.Ts, target=pn)
-    b = pn.map_throats(throats=a, target=geom2)
-    assert(sp.all(b == geom2.Ts))
-    # Check geom3
-    a = geom3.map_pores(pores=geom3.Ps, target=pn)
-    b = pn.map_pores(pores=a, target=geom3)
-    assert(sp.all(b == geom3.Ps))
-    a = geom3.map_throats(throats=geom3.Ts, target=pn)
-    b = pn.map_throats(throats=a, target=geom3)
-    assert(sp.all(b == geom3.Ts))
